Move MaatAPI schemas to a dedicated folder
