+++ crime-means-assessment/src/main/resources/schemas/maatapi/maatApiAssessmentRequest.json
@@ -79,14 +79,14 @@
       "type": "array",
       "minItems": 1,
       "items": {
-        "$ref": "common/apiAssessmentDetail.json"
+        "$ref": "../common/apiAssessmentDetail.json"
       }
     },
     "childWeightings": {
       "type": "array",
       "minItems": 1,
       "items": {
-        "$ref": "common/apiAssessmentChildWeighting.json"
+        "$ref": "../common/apiAssessmentChildWeighting.json"
       }
     },
     "dateCompleted": {
