+++ crime-means-assessment/src/main/resources/schemas/maatapi/maatApiCreateAssessment.json
