+++ crime-means-assessment/src/main/resources/schemas/maatapi/maatApiUpdateAssessment.json
