+++ crime-means-assessment/src/test/java/uk/gov/justice/laa/crime/meansassessment/builder/MeansAssessmentRequestDTOBuilderTest.java
@@ -71,7 +71,6 @@ public class MeansAssessmentRequestDTOBuilderTest {
             assertThat(resultDto.getOtherHousingNote()).isEqualTo(updateMeansAssessment.getOtherHousingNote());
             assertThat(resultDto.getInitTotalAggregatedIncome()).isEqualTo(updateMeansAssessment.getInitTotalAggregatedIncome());
             assertThat(resultDto.getFullAssessmentNotes()).isEqualTo(updateMeansAssessment.getFullAssessmentNotes());
-            assertThat(resultDto.getTimeStamp()).isEqualTo(updateMeansAssessment.getTimestamp());
         });
     }
 }
