+++ crime-means-assessment/src/test/java/uk/gov/justice/laa/crime/meansassessment/validation/validator/MeansAssessmentValidationProcessorTest.java
@@ -81,7 +81,6 @@ public class MeansAssessmentValidationProcessorTest {
         verify(initAssessmentValidator).validate(createMeansAssessmentRequest);
 
         verify(fullAssessmentValidator, never()).validate(fullAssessment);
-        verify(meansAssessmentValidationService, never()).isAssessmentModifiedByAnotherUser(createMeansAssessmentRequest);
 
         assertThat(result).isEmpty();
     }
